\name{print.logit.normal.mle}
\alias{print.logit.normal.mle}
\title{ Print summary for a Logit Normal MLE Model  }
\description{

Print formatted output of a logit normal MLE model for longitudinal binary outcome data.

}
\usage{
print.logit.normal.mle(x, digits=NULL,...)
}
\arguments{
\item{x}{
an object of class \code{logit.normal.mle}.
}
\item{digits}{
minimal number of significant digits.}
\item{...}{
further arguments passed to or from other methods.}
}
\value{
Prints a summary of class \code{logit.normal.mle}. 
}
\section{DETAILS}{
None.

}

\references{  }
\author{ Bryan Comstock, bac4@u.washington.edu }

\seealso{ \code{\link{logit.normal.mle}} }
\examples{
}
\keyword{ }
